+++ src/tool/python/code_writers.h
@@ -37,6 +37,55 @@ namespace pywinrt
         w.write("% WARNING: Please don't edit this file. It was generated by Python/WinRT\n\n", comment_marker);
     }
 
+	void write_python_setup_filenames(writer& w, std::vector<std::string> const& namespaces)
+	{
+		XLANG_ASSERT(namespaces.size() > 0);
+
+		for (auto&& ns : namespaces)
+		{
+			w.write("'%/src/py.%.cpp', ", settings.module, ns);
+		}
+
+		w.write("'%/src/_%.cpp'", settings.module, settings.module);
+	}
+
+	void write_python_import_type(writer& w, TypeDef const& type)
+	{
+		if (!is_exclusive_to(type))
+		{
+			w.write("@ = _ns_module.@\n", type.TypeName(), type.TypeName());
+		}
+	}
+
+	void write_lower_case(writer& w, std::string_view const& ns)
+	{
+		for (auto c : ns)
+		{
+			w.write(static_cast<char>(::tolower(c)));
+		}
+	}
+
+	void write_python_import_namespace(writer& w, std::string_view const& ns)
+	{
+		w.write("\ntry:\n    import %.%\nexcept:\n    pass\n", settings.module, bind<write_lower_case>(ns));
+	}
+
+	void write_python_enum(writer& w, TypeDef const& type)
+	{
+		w.write("\nclass %(enum.%):\n", type.TypeName(), is_flags_enum(type) ? "IntFlag" : "IntEnum");
+		{
+			writer::indent_guard g{ w };
+
+			for (auto&& field : type.FieldList())
+			{
+				if (auto constant = field.Constant())
+				{
+					w.write("% = %\n", bind<write_upper_snake_case>(field.Name()), *constant);
+				}
+			}
+		}
+	}
+
     void write_include(writer& w, std::string_view const& ns)
     {
         if (w.current_namespace != ns)
